requirements.txt
Flask==1.1.2
Pillow==7.2.0
numpy==1.19.1
scipy==1.5.2
opencv-python==4.4.0.40
torch==1.6.0